# api/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Duke decentralize bc",
    "version": "0.0.1",
    "title": "bc server"
  },
  "host": "localhost:3000",
  "basePath": "/api",
...
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json",
    "application/x-www-form-urlencoded",
    "application/xml",
    "text/xml"
  ],
  "produces": [
    "application/json",
    "application/xml",
    "text/xml",
    "application/javascript",
    "text/javascript"
  ],
  "paths": {
...
    "/Bank": {
      "get": {
        "tags": [
          "Bank"
        ],
        "summary": "Get all banks",
        "operationId": "getAllBanks",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Request was successful",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Bank"
              }
            }
          },
          "400": {
            "description": "Bad Request exception",
            "schema": {
              "$ref": "#/definitions/BadRequestError"
            }
          },
          "500": {
            "description": "Internal server exception",
            "schema": {
              "$ref": "#/definitions/InternalError"
            }
          }
        },
        "x-swagger-router-controller": "Bank"
      },
      "post": {
        "tags": [
          "Bank"
        ],
        "summary": "Create a new Bank",
        "operationId": "createBank",
        "parameters": [
          {
            "in": "body",
            "name": "data",
            "description": "Bank object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Bank"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request was successful",
            "schema": {
              "$ref": "#/definitions/Bank"
            }
          },
          "400": {
            "description": "Bad Request exception",
            "schema": {
              "$ref": "#/definitions/BadRequestError"
            }
          },
          "500": {
            "description": "Internal server exception",
            "schema": {
              "$ref": "#/definitions/InternalError"
            }
          }
        },
        "x-swagger-router-controller": "Bank"
      }
    },
    "/Bank/{id}": {
      "get": {
        "tags": [
          "Bank"
        ],
        "summary": "Find bank by id",
        "operationId": "findBankById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Bank object's id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Request was successful",
            "schema": {
              "$ref": "#/definitions/Bank"
            }
          },
          "400": {
            "description": "Bad Request exception",
            "schema": {
              "$ref": "#/definitions/BadRequestError"
            }
          },
          "500": {
            "description": "Internal server exception",
            "schema": {
              "$ref": "#/definitions/InternalError"
            }
          }
        },
        "x-swagger-router-controller": "Bank"
      },
      "put": {
        "tags": [
          "Bank"
        ],
        "summary": "Update bank by id",
        "operationId": "updateBank",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Bank's id",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "data",
            "description": "Bank's object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Bank"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request was successful",
            "schema": {
              "$ref": "#/definitions/Bank"
            }
          },
          "400": {
            "description": "Bad Request exception",
            "schema": {
              "$ref": "#/definitions/BadRequestError"
            }
          },
          "500": {
            "description": "Internal server exception",
            "schema": {
              "$ref": "#/definitions/InternalError"
            }
          }
        },
        "x-swagger-router-controller": "Bank"
      },
      "delete": {
        "tags": [
          "Bank"
        ],
        "summary": "Delete bank by id",
        "operationId": "deleteBank",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Bank's id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Request was successful",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "400": {
            "description": "Bad Request exception",
            "schema": {
              "$ref": "#/definitions/BadRequestError"
            }
          },
          "500": {
            "description": "Internal server exception",
            "schema": {
              "$ref": "#/definitions/InternalError"
            }
          }
        },
        "x-swagger-router-controller": "Bank"
      }
    },
...
    "/Energy": {
      "get": {
        "tags": [
          "Energy"
        ],
        "summary": "Get all energy assets",
        "operationId": "getAllEnergy",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Request was successful",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Energy"
              }
            }
          },
          "400": {
            "description": "Bad Request exception",
            "schema": {
              "$ref": "#/definitions/BadRequestError"
            }
          },
          "500": {
            "description": "Internal server exception",
            "schema": {
              "$ref": "#/definitions/InternalError"
            }
          }
        },
        "x-swagger-router-controller": "Energy"
      },
      "post": {
        "tags": [
          "Energy"
        ],
        "summary": "Create a new Energy Asset",
        "operationId": "createEnergy",
        "parameters": [
          {
            "in": "body",
            "name": "data",
            "description": "Energy object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Energy"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request was successful",
            "schema": {
              "$ref": "#/definitions/Energy"
            }
          },
          "400": {
            "description": "Bad Request exception",
            "schema": {
              "$ref": "#/definitions/BadRequestError"
            }
          },
          "500": {
            "description": "Internal server exception",
            "schema": {
              "$ref": "#/definitions/InternalError"
            }
          }
        },
        "x-swagger-router-controller": "Energy"
      }
    },
    "/Energy/{id}": {
      "get": {
        "tags": [
          "Energy"
        ],
        "summary": "Find energy asset by id",
        "operationId": "getEnergyById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Energy id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Request was successful",
            "schema": {
              "$ref": "#/definitions/Energy"
            }
          },
          "400": {
            "description": "Bad Request exception",
            "schema": {
              "$ref": "#/definitions/BadRequestError"
            }
          },
          "500": {
            "description": "Internal server exception",
            "schema": {
              "$ref": "#/definitions/InternalError"
            }
          }
        },
        "x-swagger-router-controller": "Energy"
      },
      "put": {
        "tags": [
          "Energy"
        ],
        "summary": "Update energy Asset",
        "operationId": "updateEnergy",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Energy id",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "data",
            "description": "Energy object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Energy"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request was successful",
            "schema": {
              "$ref": "#/definitions/Energy"
            }
          },
          "400": {
            "description": "Bad Request exception",
            "schema": {
              "$ref": "#/definitions/BadRequestError"
            }
          },
          "500": {
            "description": "Internal server exception",
            "schema": {
              "$ref": "#/definitions/InternalError"
            }
          }
        },
        "x-swagger-router-controller": "Energy"
      },
      "delete": {
        "tags": [
          "Energy"
        ],
        "summary": "Delete energy Asset",
        "operationId": "deleteEnergy",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Energy id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Request was successful",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "400": {
            "description": "Bad Request exception",
            "schema": {
              "$ref": "#/definitions/BadRequestError"
            }
          },
          "500": {
            "description": "Internal server exception",
            "schema": {
              "$ref": "#/definitions/InternalError"
            }
          }
        },
        "x-swagger-router-controller": "Energy"
      }
    },
    "/Cash": {
      "get": {
        "tags": [
          "Cash"
        ],
        "summary": "Get all cash assets",
        "operationId": "getAllCash",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Request was successful",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Cash"
              }
            }
          },
          "400": {
            "description": "Bad Request exception",
            "schema": {
              "$ref": "#/definitions/BadRequestError"
            }
          },
          "500": {
            "description": "Internal server exception",
            "schema": {
              "$ref": "#/definitions/InternalError"
            }
          }
        },
        "x-swagger-router-controller": "Cash"
      },
      "post": {
        "tags": [
          "Cash"
        ],
        "summary": "Create a new cash asset",
        "operationId": "createCash",
        "parameters": [
          {
            "in": "body",
            "name": "data",
            "description": "Model instance data",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Cash"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request was successful",
            "schema": {
              "$ref": "#/definitions/Cash"
            }
          },
          "400": {
            "description": "Bad Request exception",
            "schema": {
              "$ref": "#/definitions/BadRequestError"
            }
          },
          "500": {
            "description": "Internal server exception",
            "schema": {
              "$ref": "#/definitions/InternalError"
            }
          }
        },
        "x-swagger-router-controller": "Cash"
      }
    },
    "/Cash/{id}": {
      "get": {
        "tags": [
          "Cash"
        ],
        "summary": "Find cash asset by id",
        "operationId": "getCashById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Cash id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Request was successful",
            "schema": {
              "$ref": "#/definitions/Cash"
            }
          },
          "400": {
            "description": "Bad Request exception",
            "schema": {
              "$ref": "#/definitions/BadRequestError"
            }
          },
          "500": {
            "description": "Internal server exception",
            "schema": {
              "$ref": "#/definitions/InternalError"
            }
          }
        },
        "x-swagger-router-controller": "Cash"
      },
      "put": {
        "tags": [
          "Cash"
        ],
        "summary": "Update cash asset",
        "operationId": "updateCash",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Cash id",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "data",
            "description": "Cash object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Cash"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request was successful",
            "schema": {
              "$ref": "#/definitions/Cash"
            }
          },
          "400": {
            "description": "Bad Request exception",
            "schema": {
              "$ref": "#/definitions/BadRequestError"
            }
          },
          "500": {
            "description": "Internal server exception",
            "schema": {
              "$ref": "#/definitions/InternalError"
            }
          }
        },
        "x-swagger-router-controller": "Cash"
      },
      "delete": {
        "tags": [
          "Cash"
        ],
        "summary": "Delete cash by id",
        "operationId": "deleteCash",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Cash id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Request was successful",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "400": {
            "description": "Bad Request exception",
            "schema": {
              "$ref": "#/definitions/BadRequestError"
            }
          },
          "500": {
            "description": "Internal server exception",
            "schema": {
              "$ref": "#/definitions/InternalError"
            }
          }
        },
        "x-swagger-router-controller": "Cash"
      }
    },
...
  },
  "definitions": {
    "InternalError": {
      "properties": {
        "message": {
          "type": "string",
          "default": "Internal Server Error"
        }
      },
      "description": "Internal Server Error"
    },
    "BadRequestError": {
      "properties": {
        "message": {
          "type": "string",
          "default": "Bad Request Error"
        }
      },
      "description": "Bad Request Error"
    },
...
    "Bank": {
      "type": "object",
      "required": [
        "cash",
        "name",
        "tokens"
      ],
      "properties": {
        "id": {
          "type": "string",
          "description": "Primary key for struct bank"
        },
        "name": {
          "type": "string",
          "description": "Name of bank"
        },
        "tokens": {
          "type": "string",
          "description": "The identifier of an struc of tokens"
        },
        "cash": {
          "type": "string",
          "description": "The identifier of an struct of cash"
        }
      },
      "description": "A participant named Bank",
      "example": {
        "name": "name",
        "tokens": "tokens",
        "id": "id",
        "cash": "cash"
      }
    },
...
    "Energy": {
      "type": "object",
      "required": [
        "owner",
        "ownerType",
        "units",
        "value"
      ],
      "properties": {
        "id": {
          "type": "string",
          "description": "The instance identifier for this struct Energy"
        },
        "units": {
          "type": "string",
          "enum": [
            "kWh"
          ]
        },
        "value": {
          "type": "number",
          "format": "double"
        },
        "owner": {
          "type": "string"
        },
        "ownerType": {
          "type": "string",
          "enum": [
            "Resident",
            "Bank",
            "Utility"
          ]
        }
      },
      "description": "An asset named Energy",
      "example": {
        "owner": "owner",
        "ownerType": "Resident",
        "id": "id",
        "units": "kWh",
        "value": 0.8008281904610115
      }
    },
    "Cash": {
      "type": "object",
      "required": [
        "currency",
        "owner",
        "ownerType",
        "value"
      ],
      "properties": {
        "id": {
          "type": "string",
          "description": "The instance identifier for this struct Cash"
        },
        "currency": {
          "type": "string",
          "enum": [
            "USD"
          ]
        },
        "value": {
          "type": "number",
          "format": "double"
        },
        "owner": {
          "type": "string"
        },
        "ownerType": {
          "type": "string",
          "enum": [
            "Resident",
            "Bank",
            "Utility"
          ]
        }
      },
      "description": "An asset named Cash",
      "example": {
        "owner": "owner",
        "ownerType": "Resident",
        "currency": "USD",
        "id": "id",
        "value": 0.8008281904610115
      }
    },
...
  }
}